Fix the logic error for 'timeout' scenario.
Sometimes, even the job is timeout, CI would try to prepare the result
and then the script will try to get the failed cases. It does not always
make senses. Now the script would check the result. If the result is
'timeout' or there is no report, the script will not try to get the
failed cases number.